backend/requirements.txt
Django>=5.0
Pillow
djangorestframework
djangorestframework-simplejwt
gunicorn
psycopg2-binary
drf-yasg
dj-database-url
django-cors-headers